feat: 적금 상품 목록을 가져오는 기능 추가
