+++ README.md
@@ -16,7 +16,7 @@
 
 - [x] HTTP client는 "_손쉬운 API 사용을 위해 HTTP client 구현체를 미리 제공하고 있어요. `tosslib` 패키지에서 아래와 같이 import 해서 사용할 수 있어요._"를 참고해주세요.
 - [x] 적금 상품 목록 API는 상품 목록 가져오기 (`GET /api/savings-products`) 를 참고해주세요.
-- [ ]원단위를 표시할 때 콤마(,) 출력에 유의해주세요.
+- [x] 원단위를 표시할 때 콤마(,) 출력에 유의해주세요.
 
 ### 2. 저축 목표 입력 기능 만들기
