+++ src/domain/savingsCalculator/SavingsCalculatorPage.tsx
@@ -1,5 +1,5 @@
 import { savingsProducts } from 'api/savingsProducts';
-import { useEffect } from 'react';
+import { useEffect, useState } from 'react';
 import {
   Assets,
   Border,
@@ -13,8 +13,28 @@ import {
   Tab,
   TextField,
 } from 'tosslib';
+import { SavingsProduct } from 'types/savingsProduct';
 
 export function SavingsCalculatorPage() {
+  const [savingsProductList, setSavingsProductList] = useState<SavingsProduct[]>([]);
+
+  useEffect(() => {
+    async function fetchSavingsProducts() {
+      try {
+        const products = await savingsProducts();
+        setSavingsProductList(products);
+      } catch (error) {
+        if (isHttpError(error)) {
+          console.error('HTTP Error:', error.status, error.message);
+        } else {
+          console.error('Unexpected Error:', error);
+        }
+      }
+    }
+
+    fetchSavingsProducts();
+  }, []);
+
   return (
     <>
       <NavigationBar title="적금 계산기" />
@@ -44,35 +64,23 @@ export function SavingsCalculatorPage() {
         </Tab.Item>
       </Tab>
 
-      <ListRow
-        contents={
-          <ListRow.Texts
-            type="3RowTypeA"
-            top={'기본 정기적금'}
-            topProps={{ fontSize: 16, fontWeight: 'bold', color: colors.grey900 }}
-            middle={'연 이자율: 3.2%'}
-            middleProps={{ fontSize: 14, color: colors.blue600, fontWeight: 'medium' }}
-            bottom={'100,000원 ~ 500,000원 | 12개월'}
-            bottomProps={{ fontSize: 13, color: colors.grey600 }}
-          />
-        }
-        right={<Assets.Icon name="icon-check-circle-green" />}
-        onClick={() => {}}
-      />
-      <ListRow
-        contents={
-          <ListRow.Texts
-            type="3RowTypeA"
-            top={'고급 정기적금'}
-            topProps={{ fontSize: 16, fontWeight: 'bold', color: colors.grey900 }}
-            middle={'연 이자율: 2.8%'}
-            middleProps={{ fontSize: 14, color: colors.blue600, fontWeight: 'medium' }}
-            bottom={'50,000원 ~ 1,000,000원 | 24개월'}
-            bottomProps={{ fontSize: 13, color: colors.grey600 }}
-          />
-        }
-        onClick={() => {}}
-      />
+      {savingsProductList.map(product => (
+        <ListRow
+          key={product.id}
+          contents={
+            <ListRow.Texts
+              type="3RowTypeA"
+              top={product.name}
+              topProps={{ fontSize: 16, fontWeight: 'bold', color: colors.grey900 }}
+              middle={`연 이자율: ${product.annualRate}%`}
+              middleProps={{ fontSize: 14, color: colors.blue600, fontWeight: 'medium' }}
+              bottom={`${product.minMonthlyAmount.toLocaleString()}원 ~ ${product.maxMonthlyAmount.toLocaleString()}원 | ${product.availableTerms}개월`}
+              bottomProps={{ fontSize: 13, color: colors.grey600 }}
+            />
+          }
+          onClick={() => {}}
+        />
+      ))}
 
       {/* 아래는 계산 결과 탭 내용이에요. 계산 결과 탭을 구현할 때 주석을 해제해주세요. */}
       {/* <Spacing size={8} />
