+++ src/types/savingsProduct.ts
@@ -1,0 +1,8 @@
+export interface SavingsProduct {
+  id: string;
+  name: string;
+  annualRate: number;
+  minMonthlyAmount: number;
+  maxMonthlyAmount: number;
+  availableTerms: number;
+}
